requirements.txt
numpy==1.24.2
openai==0.27.2
asyncio==3.4.3
pyaudio==0.2.13
google-cloud-speech==2.18.0
google.api_core==2.11.0
pydub==0.25.1
requests==2.28.2
pyttsx3===2.90